finally figured out the code refactor problem
