+++ src/App.js
@@ -27,6 +27,8 @@ import EasySudoku from './components/puzzles-by-difficulty/easy/EasySudoku';
 import UploadForm from './components/upload-image/UploadForm';
 import UploadSudoku2 from './components/upload-image/UploadSudoku2';
 
+import puzzleHandler from './components/SudokuGame/puzzle-handler/ConstructPuzzle';
+
 const App = () => {
   const [token, setToken] = useState(localStorage.getItem('token'));
 
@@ -61,7 +63,7 @@ const App = () => {
           <Route path='/random' component={Sudoku} />
           <Route path='/medium' component={MediumSudoku} />
           <Route path='/diabolical' component={DiabolicalSudoku} />
-          <Route path='/easy' component={EasySudoku} />
+          <Route path='/easy' component={puzzleHandler} />
           <Route path='/hard' component={HardSudoku} />
           <Route path='/user-puzzles' component={ResumedPuzzle} />
           <Route path='/upload' component={UploadForm} />
