+++ src/components/Sudoku.css
@@ -1,185 +1,64 @@
-.Sudoku {
-	background-image: url("../images/Diamonds.jpg");
-	height: 100vh;
-	text-align:center;
-	font-family: 'Roboto'; font-size: 22px;
-	height: 100%;
+/* input {
+  padding: 0;
+  margin: 0;
+  height: 40px;
+  width: 40px;
+  font-size: 28px;
+  font-weight: bold;
+  border-color: grey;
+  background: #ffffff;
+} */
 
+input[type='text']:disabled {
+  background-color: #cdcac9;
 }
-
-.Board {
-	height:auto;
-	width: 30vw;
-	display: inline-block;
-	margin: auto;
-
+input[type='text']:disabled {
+  color: black;
 }
-
-input {
-	padding: 0;
-	margin: 0;
-	height: 40px;
-	width: 40px;
-	font-size: 28px;
-	font-weight: bold;
-	border-color:  grey;
-	background: #FFFFFF;
+input[type='text']:focus {
+  background: #608cff;
 }
-
-input[type="text"]:disabled{background-color: #cdcac9;}
-input[type="text"]:disabled{color: black;}
-input[type="text"]:focus{background: #608CFF;}
 
 .Square_Input {
-	padding: 0;
-	margin: 0;
-	height: 100%;
-	width: 100%;
-	
+  padding: 0;
+  margin: 0;
+  height: 100%;
+  width: 100%;
 }
 
 .Square {
-	border: 0.5px solid grey;
-	padding: 0;
-	width: 100%;
-	height: 100%;
-	text-align:center;
+  border: 0.5px solid grey;
+  padding: 0;
+  width: 100%;
+  height: 100%;
+  text-align: center;
 }
 
-.Board_RowData {
-	display: flex;
-	width: 30vw;
-	height: 7vh;
-}
+/* ---------------------------- */
 
 .SudokuButtons {
-	display: inline-flex;
-	margin-bottom:2%;
+  display: inline-flex;
+  margin-bottom: 2%;
 }
 
 .alert {
-	padding: 20px;
-	background-color: #608CFF;
-	color: black;
-	margin-bottom: 15px;
-	opacity: 1;
-	transition: opacity 0.6s; /* 600ms to fade out */
-  }
-  
+  padding: 20px;
+  background-color: #608cff;
+  color: black;
+  margin-bottom: 15px;
+  opacity: 1;
+  transition: opacity 0.6s; /* 600ms to fade out */
+}
+
 .label svg {
-	size:'30em',
-}
-
-.dark-mode .Sudoku {
-	color: #fff;
-	background-color: #313843;
-	background-image: none;
-	height: 100%;
-}
-
-.dark-mode input {
-	color: #fff;
-	background: slategrey;
-	border-color:  black;
-}
-
-.nature-mode .Sudoku {
-	background-image: url("../images/forest.png");
-	background-size: 100%;
-	color: #fff;
-	height: 100%;
-
-}
-
-.nature-mode input {
-	color: #fff;
-	background: slategrey;
-	border-color:  black;
-}
-
-.nature-mode input[type="text"]:enabled {
-	opacity:.8 !important;
-	background-color:#acd052;
-	color:black;
-}
-
-.nature-mode input[type="text"]:disabled {
-	opacity:.8 !important;
-	background-color:#4c712f;
-	color:white;
-}
-
-.nature-mode input[type="text"]:focus {
-	background: #f5f853;
-}
-
-.city-mode .Sudoku {
-	color: #fff;
-	background-image: url("../images/city.png");
-	background-size: 100%;
-	height: 100%;
-
-}
-
-.city-mode input {
-	color: #fff;
-	background: slategrey;
-	border-color:  black;
-}
-
-.city-mode input[type="text"]:enabled {
-	opacity:.8 !important;
-	background-color:#cdd4d9;
-	color:black;
-
-}
-
-.city-mode input[type="text"]:disabled {
-	opacity:.8 !important;
-	background-color:black;
-	color:white;
-}
-
-.city-mode input[type="text"]:focus {
-	background: #608cff;
-}
-
-.sky-mode .Sudoku {
-	color: black;
-	background-image: url("../images/sky.png");
-	background-size: 100%;
-	height: 100%;
-
-}
-
-.sky-mode input {
-	color: #fff;
-	background: slategrey;
-	border-color:  black;
-}
-
-.sky-mode input[type="text"]:enabled {
-	opacity:.8 !important;
-	background-color:#b2ddf3;
-	color:black;
-
-}
-
-.sky-mode input[type="text"]:disabled {
-	opacity:.8 !important;
-	background-color:#163a4e;
-	color:white;
-}
-
-.sky-mode input[type="text"]:focus {
-	background: #45bcf5;
+  size: '30em';
 }
 
 @media only screen and (max-width: 600px) {
-	.Board {
-	  width: 90%;
-	}
-	.Board_RowData {
-		width: 100%;
-	}
+  .Board {
+    width: 90%;
   }
+  .Board_RowData {
+    width: 100%;
+  }
+}
