+++ src/components/SudokuGame/puzzle-handler/functions/getDeepCopyOfArray.js
@@ -1,0 +1,6 @@
+import React from 'react';
+import { stringify } from './stringify';
+
+export function getDeepCopyOfArray(arr) {
+  return JSON.parse(JSON.stringify(arr));
+}
