+++ src/components/SudokuGame/puzzle-handler/functions/stringify.js
@@ -1,0 +1,5 @@
+import React from 'react';
+
+export function stringify(num1, num2) {
+  return num1 + '' + num2;
+}
