+++ src/components/SudokuGame/puzzle-handler/grid-axios-call/9x9.js
@@ -2,7 +2,7 @@ import axios from 'axios';
 
 export const Get9x9 = () =>
   axios
-    .get('https://omega2020.herokuapp.com/puzzle/9x9')
+    .get('https://omega2020.herokuapp.com/puzzle/gentle')
     .then((res) => {
       console.log('AXIOS', res.data);
       return res.data;
