--- src/tests/userstest.js
@@ -1,15 +1,0 @@
-// import axios from 'axios';
-// import user from '../Login.js';
-
-// jest.mock('axios');
-
-// test('should fetch users', () => {
-//   const user = [{name: 'bob', password: 'bob'}];
-//   const resp = {data: user};
-//   axios.get.mockResolvedValue(resp);
-
-//   // or you could use the following depending on your use case:
-//   // axios.get.mockImplementation(() => Promise.resolve(resp))
-
-//   return user.all().then(data => expect(data).toEqual(users));
-// });
